Create EpisodeResolver
