+++ src/modules/watchlist/resolvers/index.ts
@@ -1,1 +1,2 @@
 export { WatchlistResolver } from './watchlist.resolvers';
+export { EpisodeResolver } from './episode.resolvers';
