+++ src/modules/watchlist/resolvers/inputs/index.ts
@@ -1,1 +1,2 @@
 export { UpsertWatchlistInput } from './upsert-watchlist.input';
+export { UpsertEpisodeInput } from './upsert-episode.input';
