+++ src/modules/watchlist/resolvers/inputs/upsert-episode.input.ts
@@ -1,0 +1,10 @@
+import { Field, InputType, Int } from '@nestjs/graphql';
+
+@InputType()
+export class UpsertEpisodeInput {
+  @Field(() => Int)
+  episodeId: number;
+
+  @Field(() => Boolean)
+  isWatched: boolean;
+}
