+++ src/modules/watchlist/resolvers/watchlist.resolvers.ts
@@ -1,24 +1,15 @@
 import 'reflect-metadata';
-import { Args, Context, Info, Mutation, Query } from '@nestjs/graphql';
+import { Args, Context, Mutation, Query } from '@nestjs/graphql';
 import { Injectable, UseGuards } from '@nestjs/common';
-import { when } from 'rambda';
-import { fieldsMap } from 'graphql-fields-list';
-import { GraphQLResolveInfo } from 'graphql';
 import { GraphqlJwtAuthGuard } from '../../auth/guards';
 import { RequestWithUser } from '../../auth/interfaces';
-import { Episode } from '../../show/entities/episode';
 import { Watchlist } from '../entities';
-import { EpisodeService, WatchlistService } from '../services';
-import { ShowService } from '../../show/services';
+import { WatchlistService } from '../services';
 import { UpsertWatchlistInput } from './inputs';
 
 @Injectable()
 export class WatchlistResolver {
-  constructor(
-    private readonly watchlistService: WatchlistService,
-    private readonly episodeService: EpisodeService,
-    private readonly showService: ShowService,
-  ) {}
+  constructor(private readonly watchlistService: WatchlistService) {}
 
   @Mutation(() => Watchlist)
   @UseGuards(GraphqlJwtAuthGuard)
@@ -36,20 +27,5 @@ export class WatchlistResolver {
   @UseGuards(GraphqlJwtAuthGuard)
   async getWatchlist(@Context() { req: { user } }: { req: RequestWithUser }) {
     return await this.watchlistService.listByUserId(user.id);
-  }
-
-  @Query(() => [Episode])
-  @UseGuards(GraphqlJwtAuthGuard)
-  async listUpNext(
-    @Info() info: GraphQLResolveInfo,
-    @Context() { req: { user } }: { req: RequestWithUser },
-  ) {
-    const userId = user.id;
-    const fields = fieldsMap(info);
-
-    return this.watchlistService
-      .listUpNext(userId)
-      .then((data) => data.map((item) => ({ ...item, isWatched: false })))
-      .then(when(() => 'show' in fields, this.showService.linkShows));
   }
 }
